+++ nautilus_trader/accounting/accounts/base.pyx
@@ -21,9 +21,9 @@ from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.core.rust.model cimport AccountType
 from nautilus_trader.core.rust.model cimport OrderSide
 from nautilus_trader.model.events.account cimport AccountState
+from nautilus_trader.model.functions cimport account_type_to_str
 from nautilus_trader.model.instruments.base cimport Instrument
 from nautilus_trader.model.objects cimport AccountBalance
-from nautilus_trader.model.objects cimport account_type_to_str
 
 
 cdef class Account: