+++ RELEASES.md
@@ -5,6 +5,7 @@ Released on TBD (UTC).
 ### Enhancements
 - Rewrote Interactive Brokers integration documentation, many thanks @benjaminsingleton
 - Added `HistoricInteractiveBrokerClient`, thanks @benjaminsingleton and @limx0
+- Added `DataEngineConfig.time_bars_interval_type` (determines the type of interval used for time aggregation `left-open` or `right-open`)
 - Added `LoggingConfig.log_colors` to optionally use ANSI codes to produce colored logs (default true to retain current behavior)
 - Added `WilderMovingAverage` in Rust, thanks @ayush-sb
 - Added all common identifier generators in Rust, thanks @filipmacek